remove unnecessary interfaces
https://softwareengineering.stackexchange.com/questions/150045/what-is-the-point-of-having-every-service-class-have-an-interface
https://www.adam-bien.com/roller/abien/entry/service_s_new_serviceimpl_why
