+++ src/main/java/cz/forgottenempire/arma3servergui/modpreset/ModPresetsService.java
@@ -4,22 +4,48 @@ import cz.forgottenempire.arma3servergui.common.ServerType;
 import cz.forgottenempire.arma3servergui.workshop.WorkshopMod;
 import java.util.Collection;
 import java.util.Optional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-public interface ModPresetsService {
+@Service
+public class ModPresetsService {
 
-    Collection<ModPreset> getAllPresets();
+    private final ModPresetsRepository presetsRepository;
 
-    Collection<ModPreset> getAllPresetsForServer(ServerType serverType);
+    @Autowired
+    public ModPresetsService(ModPresetsRepository presetsRepository) {
+        this.presetsRepository = presetsRepository;
+    }
 
-    Collection<ModPreset> getAllPresetsContainingMod(WorkshopMod mod);
+    public Collection<ModPreset> getAllPresets() {
+        return presetsRepository.findAll();
+    }
 
-    Optional<ModPreset> getModPreset(Long id);
+    public Collection<ModPreset> getAllPresetsForServer(ServerType serverType) {
+        return presetsRepository.getAllByType(serverType);
+    }
 
-    ModPreset savePreset(ModPreset preset);
+    public Collection<ModPreset> getAllPresetsContainingMod(WorkshopMod mod) {
+        return presetsRepository.getAllByModsContaining(mod);
+    }
 
-    ModPreset updatePreset(ModPreset preset);
+    public Optional<ModPreset> getModPreset(Long id) {
+        return presetsRepository.findById(id);
+    }
 
-    void deletePreset(Long preset);
+    public ModPreset savePreset(ModPreset preset) {
+        return presetsRepository.save(preset);
+    }
 
-    boolean presetWithNameExists(String name);
+    public ModPreset updatePreset(ModPreset preset) {
+        return savePreset(preset);
+    }
+
+    public void deletePreset(Long preset) {
+        presetsRepository.deleteById(preset);
+    }
+
+    public boolean presetWithNameExists(String name) {
+        return presetsRepository.existsByName(name);
+    }
 }
