--- src/main/java/cz/forgottenempire/arma3servergui/modpreset/ModPresetsServiceImpl.java
@@ -1,59 +1,0 @@
-package cz.forgottenempire.arma3servergui.modpreset;
-
-import cz.forgottenempire.arma3servergui.common.ServerType;
-import cz.forgottenempire.arma3servergui.workshop.WorkshopMod;
-import java.util.Collection;
-import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-class ModPresetsServiceImpl implements ModPresetsService {
-
-    private final ModPresetsRepository presetsRepository;
-
-    @Autowired
-    public ModPresetsServiceImpl(ModPresetsRepository presetsRepository) {
-        this.presetsRepository = presetsRepository;
-    }
-
-    @Override
-    public Collection<ModPreset> getAllPresets() {
-        return presetsRepository.findAll();
-    }
-
-    @Override
-    public Collection<ModPreset> getAllPresetsForServer(ServerType serverType) {
-        return presetsRepository.getAllByType(serverType);
-    }
-
-    @Override
-    public Collection<ModPreset> getAllPresetsContainingMod(WorkshopMod mod) {
-        return presetsRepository.getAllByModsContaining(mod);
-    }
-
-    @Override
-    public Optional<ModPreset> getModPreset(Long id) {
-        return presetsRepository.findById(id);
-    }
-
-    @Override
-    public ModPreset savePreset(ModPreset preset) {
-        return presetsRepository.save(preset);
-    }
-
-    @Override
-    public ModPreset updatePreset(ModPreset preset) {
-        return savePreset(preset);
-    }
-
-    @Override
-    public void deletePreset(Long preset) {
-        presetsRepository.deleteById(preset);
-    }
-
-    @Override
-    public boolean presetWithNameExists(String name) {
-        return presetsRepository.existsByName(name);
-    }
-}
