+++ src/main/java/cz/forgottenempire/arma3servergui/steamauth/SteamAuthService.java
@@ -1,10 +1,39 @@
 package cz.forgottenempire.arma3servergui.steamauth;
 
 import cz.forgottenempire.arma3servergui.workshop.SteamAuthDto;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-public interface SteamAuthService {
+@Service
+@Slf4j
+public class SteamAuthService {
 
-    SteamAuth getAuthAccount();
+    private final SteamAuthRepository authRepository;
 
-    void setAuthAccount(SteamAuthDto auth);
+    @Autowired
+    public SteamAuthService(SteamAuthRepository authRepository) {
+        this.authRepository = authRepository;
+    }
+
+    public SteamAuth getAuthAccount() {
+        return authRepository.findAll().stream().findFirst().orElse(new SteamAuth());
+    }
+
+    public void setAuthAccount(SteamAuthDto auth) {
+        log.info("Setting new Steam Auth account: username {}, token {}",
+                auth.getUsername(), auth.getSteamGuardToken());
+        SteamAuth persistedAuth = getAuthAccount();
+        populateAuth(auth, persistedAuth);
+        authRepository.save(persistedAuth);
+    }
+
+    private void populateAuth(SteamAuthDto auth, SteamAuth persistedAuth) {
+        persistedAuth.setUsername(auth.getUsername());
+        persistedAuth.setSteamGuardToken(auth.getSteamGuardToken());
+        if (StringUtils.isNotBlank(auth.getPassword())) {
+            persistedAuth.setPassword(auth.getPassword());
+        }
+    }
 }
